added mappings, new BLL logic

+++ Domain/Domain.App/Additional.cs
@@ -5,20 +5,20 @@ namespace Domain.App
 {
     public class Additional : DomainEntityIdMetadataUser<AppUser>
     {
-        public int Additional1 { get; set; }
+        public double Additional1 { get; set; }
         
-        public int Additional2 { get; set; }
+        public double Additional2 { get; set; }
         
-        public int Additional3 { get; set; }
+        public double Additional3 { get; set; }
         
-        public int Additional4 { get; set; }
+        public double Additional4 { get; set; }
         
-        public int Additional5 { get; set; }
+        public double Additional5 { get; set; }
         
-        public int Additional6 { get; set; }
+        public double Additional6 { get; set; }
         
-        public int Additional7 { get; set; }
+        public double Additional7 { get; set; }
         
-        public int Additional8 { get; set; }
+        public double Additional8 { get; set; }
     }
 }
